The s option now gives a percentage of the size of the input

+++ util/ego/il/il1_aux.c
@@ -164,12 +164,12 @@ short looplevel(b)
 
 
 
-short proclength(p)
+int proclength(p)
 	proc_p p;
 {
 	/* count the number of EM instructions of p */
 
-	register short cnt;
+	register int cnt;
 	register bblock_p b;
 	register line_p l;
